vharness: suppress instructions annulled by exceptions [ARCHSIM-2]
Also fix the return type of virt_machine_get_pending_exception() which
need to be signed as only < 0 signals no exception.

+++ riscv_machine.c
@@ -957,7 +957,7 @@ void riscv_set_reg(RISCVCPUState *s, int rn, uint64_t val);
 void riscv_dump_regs(RISCVCPUState *s);
 int riscv_read_insn(RISCVCPUState *s, uintptr_t *pmem_addend, uint64_t addr);
 int riscv_read_u64(RISCVCPUState *s, uint64_t *data, uint64_t addr);
-uint64_t  riscv_get_pending_exception(RISCVCPUState *s);
+int riscv_get_pending_exception(RISCVCPUState *s);
 
 void virt_machine_set_pc(VirtMachine *m, uint64_t pc)
 {
@@ -977,7 +977,7 @@ uint64_t  virt_machine_get_pc(VirtMachine *m)
   return riscv_get_pc(s->cpu_state);
 }
 
-uint64_t  virt_machine_get_pending_exception(VirtMachine *m)
+int virt_machine_get_pending_exception(VirtMachine *m)
 {
   RISCVMachine *s = (RISCVMachine *)m;
   return riscv_get_pending_exception(s->cpu_state);
